Finalize split of audiomgr and audiosfx
